Separate NRTM query tests
